Added Material UI  AppMUI.js
Replace App.js into AppMUI.js

Material UI.
While Bootstrap is a popular CSS library, it does not allow us implement styling in a manner consistent with the component model of React. The Material UI framework (see here) is a much better fit. We will refactor our app to use it , however, for ease of comparison we will implement the components in new files.

Create the file src/AppMUI.js